+++ core/DOMDocumentPlus.php
@@ -27,6 +27,32 @@ class DOMDocumentPlus extends DOMDocument {
   }
 
   /**
+   * @param string $eName
+   * @param string $aMatch
+   * @param string $to
+   * @return DOMElementPlus|null
+   */
+  public function matchElement ($eName, $aMatch, $to) {
+    $lastMatch = null;
+    /** @var DOMElementPlus $element */
+    foreach ($this->getElementsByTagName($eName) as $element) {
+      if ($element->hasAttribute($aMatch)) {
+        $d = $element->getAttribute($aMatch);
+        if (!preg_match("/^[a-z0-9.*-]+$/", $d)) {
+          Logger::user_error(sprintf(_("Invalid attribute %s value '%s'"), $aMatch, $d));
+          continue;
+        }
+        $pattern = str_replace([".", "*"], ["\.", "[a-z0-9-]+"], $d);
+        if (!preg_match("/^$pattern$/", $to)) {
+          continue;
+        }
+      }
+      $lastMatch = $element;
+    }
+    return $lastMatch;
+  }
+
+  /**
    * @param string $name
    * @param string|null $value
    * @return DOMElementPlus|DOMElement